+++ extensions/word-search/src/word-adjective-search.tsx
@@ -1,6 +1,10 @@
-import { SearchType } from "./types";
-import SearchResults from "./search-results";
+import { SearchType } from "@/types";
 
-export default function SearchRhyme() {
-  return SearchResults(SearchType.ADJECTIVE, "Search for adjectives that describe a word");
+import { useClipboardSetting } from "@/hooks/use-settings";
+
+import SearchResults from "@/components/SearchResults";
+
+export default function SearchAdjective() {
+  const useClipboard = useClipboardSetting();
+  return SearchResults(SearchType.ADJECTIVE, "Search for adjectives that describe a word", useClipboard);
 }