+++ extensions/word-search/src/word-antonym-search.tsx
@@ -1,6 +1,10 @@
-import { SearchType } from "./types";
-import SearchResults from "./search-results";
+import { SearchType } from "@/types";
+
+import { useClipboardSetting } from "@/hooks/use-settings";
+
+import SearchResults from "@/components/SearchResults";
 
 export default function SearchAntonyms() {
-  return SearchResults(SearchType.ANTONYM, "Search for antonyms");
+  const useClipboard = useClipboardSetting<Preferences.WordAntonymSearch>();
+  return SearchResults(SearchType.ANTONYM, "Search for antonyms", useClipboard);
 }