+++ extensions/word-search/src/word-rhyme-search.tsx
@@ -1,6 +1,10 @@
-import { SearchType } from "./types";
-import SearchResults from "./search-results";
+import { SearchType } from "@/types";
+
+import { useClipboardSetting } from "@/hooks/use-settings";
+
+import SearchResults from "@/components/SearchResults";
 
 export default function SearchRhyme() {
-  return SearchResults(SearchType.RHYME, "Search for rhymes");
+  const useClipboard = useClipboardSetting<Preferences.WordRhymeSearch>();
+  return SearchResults(SearchType.RHYME, "Search for rhymes", useClipboard);
 }