+++ extensions/word-search/src/word-spell.tsx
@@ -1,6 +1,10 @@
-import { SearchType } from "./types";
-import SearchResults from "./search-results";
+import { SearchType } from "@/types";
 
-export default function SearchAntonyms() {
-  return SearchResults(SearchType.MISSING_LETTERS, "Spell a word");
+import { useClipboardSetting } from "@/hooks/use-settings";
+
+import SearchResults from "@/components/SearchResults";
+
+export default function SearchSpelling() {
+  const useClipboard = useClipboardSetting<Preferences.WordSpell>();
+  return SearchResults(SearchType.MISSING_LETTERS, "Spell a word", useClipboard);
 }