+++ extensions/word-search/src/word-synonym-search.tsx
@@ -1,6 +1,10 @@
-import { SearchType } from "./types";
-import SearchResults from "./search-results";
+import { SearchType } from "@/types";
+
+import { useClipboardSetting } from "@/hooks/use-settings";
+
+import SearchResults from "@/components/SearchResults";
 
 export default function SearchSynonym() {
-  return SearchResults(SearchType.SYNONYM, "Search for synonyms");
+  const useClipboard = useClipboardSetting<Preferences.WordSynonymSearch>();
+  return SearchResults(SearchType.SYNONYM, "Search for synonyms", useClipboard);
 }